Should use `NSUInteger` instead of `int` in the API

+++ ResearchKit/Common/ORKOrderedTask.h
@@ -431,7 +431,7 @@ typedef NS_OPTIONS(NSUInteger, ORKPredefinedTaskOption) {
 
 + (ORKOrderedTask *)towerOfHanoiTaskWithIdentifier:(NSString *)identifier
                             intendedUseDescription:(nullable NSString *)intendedUseDescription
-                                     numberOfDisks:(int)numberOfDisks
+                                     numberOfDisks:(NSUInteger)numberOfDisks
                                            options:(ORKPredefinedTaskOption)options;
 
 @end
